+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.5.0] 2021-05-31
+### Changed
+- Updated Python dependencies (`requirements.txt`)
+### Fixed
+- Fixed scraping of CoinMarketCap exchange rates
+### Added
+- Added --concurrency command-line argument (#17)
+
 ## [0.4.5] - 2020-11-16
 ### Changed
 - Upgrade to BlockSci v0.7.0 (requires reparsing)